update install and uninstall scripts, update bash_profile and vimrc to support that

+++ install/uninstall.sh
@@ -1,4 +1,12 @@
-rm -rf $HOME/.cbash
-rm -rf $HOME/.bash_profile
-mv $HOME/.bash_profile.bk $HOME/.bash_profile
+INSTALLATION_FOLDER=.cbash-gurinderhans
+WD=$HOME/$INSTALLATION_FOLDER
+rm -rf $WD
+
+if [ -f $HOME/.bash_profile ]; then
+	rm -rf $HOME/.bash_profile
+fi
+
+if [ -f $HOME/.bash_profile.bk ]; then
+	mv $HOME/.bash_profile.bk $HOME/.bash_profile
+fi
